Reorganize pong module architecture
