+++ src/backend/src/pong/class/game-map/game-map.ts
