+++ src/backend/src/pong/pong.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
-import { GameGateway } from './game/game.gateway';
-import { GameService } from './game/game.service';
-import { MatchmakingGateway } from './matchmaking/matchmaking.gateway';
-import { MatchmakingService } from './matchmaking/matchmaking.service';
+import { GameGateway } from './gateway/game/game.gateway';
+import { GameGatewayService } from './gateway/game/game-gateway.service';
+import { MatchmakingGateway } from './gateway/matchmaking/matchmaking.gateway';
+import { MatchmakingService } from './gateway/matchmaking/matchmaking-gateway.service';
+import { GameCoreService } from './service/game-core/game-core.service';
 
 @Module({
   controllers: [],
   imports: [JwtModule],
-  providers: [GameGateway, GameService, MatchmakingGateway, MatchmakingService],
+  providers: [GameCoreService, GameGateway, GameGatewayService, MatchmakingGateway, MatchmakingService],
 })
 export class PongModule {}
